fix: Testing enviroment and docs fixings
