+++ package.json
@@ -108,6 +108,7 @@
     "@types/mocha": "^7.0.2",
     "@types/react": "^16.9.34",
     "@types/react-dom": "^16.9.6",
+    "@types/react-test-renderer": "^16.9.2",
     "and": "0.0.3",
     "chromedriver": "^81.0.0",
     "electron-squirrel-startup": "^1.0.0",
